Use hand-crafted RegExp instead of a generated parser
According to the newly-added benchmark, this is between 414x and 627x faster depending on the run. (This measurement was done after the return type conversion mentioned below, so it is apples-to-apples.)

The RegExp is basically unmaintainable, and ideally we would generate it in some way. However in practice the XML specification never changes so we shouldn't need to maintain it.

As part of this, change the API to just return a boolean, instead of the { success, error } object.